remove dead code

+++ src/arraytools.h
@@ -2302,7 +2302,7 @@ public:
         {
             int i = A.n_columns;
             if ( writearrays ) {
-                if ( i < afiles.size () && i>=0 ) {
+                if ( i < (int) afiles.size () && i>=0 ) {
                     afiles[i]->append_array ( A );
                 } else {
                     fprintf ( stderr,
